requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
distro==1.9.0
fastapi==0.135.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
numpy==2.4.2
openai==2.21.0
pyaec==1.0.1
PyAudio==0.2.14
pydantic==2.12.5
pydantic_core==2.41.5
sniffio==1.3.1
speexdsp-ns==0.1.2
starlette==0.52.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
webrtc_noise_gain==1.2.5
websocket-client==1.9.0
websockets>=13.0
twilio>=9.0
python-dotenv>=1.0
requests>=2.31
